Add primitives for claims pallet
