+++ primitives/claims/Cargo.toml
@@ -1,8 +1,32 @@
 [package]
-name = "claims"
+name = "claims-primitives"
 version = "0.1.0"
+authors = ["Integritee AG"]
+homepage = "https://integritee.network/"
+repository = "https://github.com/integritee-network/pallets/"
+license = "Apache-2.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+resolver = "2"
 
 [dependencies]
+scale-info = { version = "1.0", default-features = false, features = ["derive"] }
+parity-scale-codec = { version = "2.3.1", default-features = false, features = ["derive"] }
+rustc-hex = { version = "2.1.0", default-features = false }
+serde = { version = "1.0.130", default-features = false }
+
+sp-std = { package = "sp-std", git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+sp-io = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
+
+[features]
+default = ["std"]
+no_std = []
+std = [
+    "parity-scale-codec/std",
+    "scale-info/std",
+    "rustc-hex/std",
+    "sp-std/std",
+    "sp-io/std",
+    "sp-runtime/std",
+    "serde/std",
+]
